Add definition for WRITE_TIMEOUT_MS and WAIT_TIME_MS static const attributes
See https://stackoverflow.com/questions/16957458/static-const-in-c-class-undefined-reference/16957554 for the extendend explanation.
Note that this constant may be accessed by reference, depending if standalone ASIO is used, or depending on the specific Boost::Asio version.
In our case, we experienced this problem only when compiling in Debug, and when using standalone Asio or Boost 1.69, while it was not happening
when using Boost 1.65.